chore: update database

+++ src/CFMS.Application/Features/UserFeat/Delete/DeleteUserCommandHandler.cs
@@ -30,7 +30,7 @@ namespace CFMS.Application.Features.UserFeat.Delete
                 var existUser = _unitOfWork.UserRepository.GetByID(request.UserId);
                 if (existUser == null) return BaseResponse<bool>.FailureResponse("Người dùng không tồn tại");
                 if (existUser.Status.Equals(UserStatus.Inactive.ToString())) return BaseResponse<bool>.FailureResponse("Người dùng đã bị xoá");
-                existUser.Status = UserStatus.Inactive;
+                existUser.Status = (int)UserStatus.Inactive;
                 _unitOfWork.UserRepository.Update(existUser);
 
                 var tokens = _unitOfWork.RevokedTokenRepository.Get(
